requirements.txt
playwright>=1.40.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
anthropic>=0.40.0
apscheduler>=3.10.0
aiofiles>=24.1.0
markdown>=3.5.0
pyyaml>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
